dev -> main 최신화

+++ domains/admin/admin.service.js
@@ -33,8 +33,8 @@ export const createRoomsService = async (body, userId) => {
   try {
     if (!body.room_image) throw new Error("공지방 생성을 위한 프로필 이미지가 필요합니다.");
     const roomInviteUrl = createShortUUID();
-    const roomData = await createRoomsDao(body, userId, roomInviteUrl);
-    return createRoomsDTO(roomData);
+    const createRoomsData = await createRoomsDao(body, userId, roomInviteUrl);
+    return createRoomsDTO(createRoomsData);
   } catch (error) {
     console.error("공지방 생성하기 에러:", error);
     throw error;
@@ -183,6 +183,7 @@ export const userInviteService = async (roomId) => {
 
 export const deleteUserService = async (body) => {
   try {
+    if(!body.userId) throw new Error("강퇴할 User의 ID가 필요합니다.");
     const result = await deleteUserDao(body);
     if (result == -1) throw new Error("공지방에 유저가 없습니다.");
     return result;
